Updated CSS Mobile First
Removed a lot of the CSS intended for the websites desktop layout.

By rewriting the website mobile first, a fully fleshed out and working mobile website can be built up to support desktops as opposed to starting with a mobile compatible website, structuring it for desktop,  and then struggling to make it mobile compatible again.