Updated styles for image carousel
Updated the images in the image carousel to just center and just edge out the adjacent images to only see one at a time.

Also added options to the carousel to remove the next/previous buttons, allow for infinite scrolling and provide better reading for how many slide there are.