Finalizing before uploading
The layout has been more or less finalized for now. There are a couple more things I would like to try, but I also want to get it posted. As it is now, I'm very happy with the layout and how it changes when scaling from mobile to desktop.

Added proper images for the two projects.

Minor style updates including using ::after pseudo elements to toggle the box shadow on .nav-links to fade in instead of just appearing.