Updated About, Twitter link, font-sizes, and nav-link spacing
The nav-link button was cutting off part of my name at the top. That's been adjusted with some padding-top on the side-bar.

The font-size on xl-size screens was a little too big at 1.5rem (24px) and was adjusted to a flat 18px.

Updated my title to KamenRiderDev as I'm strongly leaning on that for my website title.

Updated my about summary a little to clarify some information. Also updated my Twitter link and handle to KamenRiderDev from KR_Progriser.