Updated projects section to grid
Updated the project section to now be based off grid instead of flex. This gives more flexibility in changing the layout between smaller mobile screens and desktop.

Adjusted the size html font size up to 18px on mobile, up to 1.5rem on tablets, and back down to 1rem on desktop.

Adjusted the max size of the headshot class to the max value of either 40% or 281, half the images pixel width.

Adjusted the bullet points of the About section's skills lists. They should now stay centered in the list as the screen sizes grow.

Removed the min-width: 992px media query.

Now that not as many items are based off the rem values, there wasn't a need for that many adjustments in the media queries.