Updated about and ul::after
Updated about section's container with a additional "about-flex" class to give the section a flexbox container to better arrange everything when scaling up to desktop views.

Combined the skills lists into a single list and changed the ul display to grid to give better control on the positioning and sizing of the list items.

Added "headshot-wrapper" around "headshot" to help with positioning in the flexbox.

Updated the arrow created in ul li::after to display: inline-block from position: absolute.

Removed the paragraph wrapping the list item contents in the project tabs.